Deprecate Head and Foot. Allow Body variation for own each
