--- addon/components/yeti-table/body/row/template.hbs
@@ -1,6 +1,0 @@
-{{!-- template-lint-disable no-invalid-interactive --}}
-<tr class="{{@theme.tbodyRow}} {{@theme.row}}" {{on "click" this.handleClick}} role={{if @onClick "button"}} ...attributes>
-  {{yield (hash
-    cell=(component "yeti-table/body/row/cell" theme=@theme parent=this)
-  )}}
-</tr>
