--- addon/components/yeti-table/foot/row/cell/component.js
@@ -1,53 +1,0 @@
-import Component from '@ember/component';
-
-import { tagName, layout } from '@ember-decorators/component';
-import { reads } from '@ember/object/computed';
-
-import template from './template';
-
-/**
-  Renders a `<td>` element and yields for the developer to supply content.
-
-  ```hbs
-  <table.foot as |foot|>
-    <foot.row as |row|>
-      <row.cell>
-        Footer content
-      </row.cell>
-    </foot.row>
-  </table.head>
-  ```
-
- */
-@tagName('')
-@layout(template)
-class FootCell extends Component {
-  theme;
-
-  parent;
-
-  /**
-   * Controls the visibility of the current cell. Keep in mind that this property
-   * won't just hide the column using css. The DOM for the column will be removed.
-   * Defaults to the `visible` argument of the corresponding column.
-   */
-  @reads('column.visible')
-  visible;
-
-  init() {
-    super.init(...arguments);
-
-    if (this.get('parent')) {
-      this.get('parent').registerCell(this);
-    }
-  }
-
-  willDestroyElement() {
-    super.willDestroyElement(...arguments);
-    if (this.get('parent')) {
-      this.get('parent').unregisterCell(this);
-    }
-  }
-}
-
-export default FootCell;
