--- addon/components/yeti-table/foot/row/template.hbs
@@ -1,5 +1,0 @@
-<tr class="{{@class}} {{@theme.tfootRow}} {{@theme.row}}" ...attributes>
-  {{yield (hash
-    cell=(component "yeti-table/foot/row/cell" theme=@theme parent=this)
-  )}}
-</tr>
