--- addon/components/yeti-table/foot/template.hbs
@@ -1,5 +1,0 @@
-<tfoot class={{@theme.tfoot}} ...attributes>
-  {{yield (hash
-    row=(component "yeti-table/foot/row" columns=@columns theme=@theme parent=@parent)
-  )}}
-</tfoot>
