--- addon/components/yeti-table/head/template.hbs
@@ -1,5 +1,0 @@
-<thead class={{@theme.thead}} ...attributes>
-  {{yield (hash
-    row=(component "yeti-table/head/row" sortable=@sortable sortSequence=@sortSequence onColumnClick=@onColumnClick columns=@columns theme=@theme parent=@parent)
-  )}}
-</thead>
