+++ addon/components/yeti-table/header/template.hbs
@@ -1,7 +1,7 @@
 <thead class={{@theme.thead}} ...attributes>
   <tr class="{{@trClass}} {{@theme.theadRow}} {{@theme.row}}">
     {{yield (hash
-      column=(component "yeti-table/head/row/column" sortable=@sortable sortSequence=@sortSequence onClick=(fn this.onColumnClickHeader) parent=@parent theme=@theme)
+      column=(component "ember-yeti-table@yeti-table/thead/row/column/component" sortable=@sortable sortSequence=@sortSequence onClick=(fn this.onColumnClickHeader) parent=@parent theme=@theme)
     )}}
   </tr>
 </thead>
