+++ addon/components/yeti-table/tbody/template.hbs
@@ -1,0 +1,6 @@
+<tbody class={{@theme.tbody}} ...attributes>
+  {{yield
+    (hash row=(component "ember-yeti-table@yeti-table/tbody/row/component" theme=@theme onClick=this.onRowClick columns=@columns))
+    @data
+  }}
+</tbody>
