+++ addon/components/yeti-table/tfoot/row/cell/template.hbs
