+++ addon/components/yeti-table/tfoot/row/template.hbs
@@ -1,0 +1,5 @@
+<tr class="{{@class}} {{@theme.tfootRow}} {{@theme.row}}" ...attributes>
+  {{yield (hash
+    cell=(component "ember-yeti-table@yeti-table/tfoot/row/cell/component" theme=@theme parent=this)
+  )}}
+</tr>
