+++ addon/components/yeti-table/thead/row/cell/template.hbs
