--- app/components/yeti-table/body.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-yeti-table/components/yeti-table/body/component';
