--- app/components/yeti-table/foot.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-yeti-table/components/yeti-table/foot/component';
