--- app/components/yeti-table/head.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-yeti-table/components/yeti-table/head/component';
