--- app/components/yeti-table/head/row/cell.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-yeti-table/components/yeti-table/head/row/cell/component';
