--- app/components/yeti-table/header.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-yeti-table/components/yeti-table/header/component';
