--- app/components/yeti-table/pagination.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-yeti-table/components/yeti-table/pagination/component';
