--- app/components/yeti-table/table.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-yeti-table/components/yeti-table/table/component';
